Allow sending media metadata.
This defines the structs, and an encoder, but not yet a decoder. This
allows sending metadata such as track title and album art with media.

Most of the fields map directly to those as specified in [1], but some
fields have been merged to take advantage of Rust's type system. For
example, it does not make sense to provide a latitude for a photograph,
but not a longitude. For clarity, the subtitle field of a TV show was
renamed to episode_title.

[1]: https://developers.google.com/cast/docs/reference/messages